Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Missing Tests !!

Tests are available but not in this commit. We tried to minimise the size of the commit to hopefully make the initial review a bit easier. 

If you prefer we can abandon this change and upstream the entire seed code contribution in one massive commit - then all tests etc would be included - but very difficult to review. 

Please let us know which is better for you. :-)

Patch-set: 4
Reviewer: Gerrit User 5519 <5519@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>