Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Any idea why this fix works? 
> I think this worked fine before?

The test has not been included in the automatic tests, so it has probably been run by me locally. 
It is a new test for the new V2 of the NBI. 
The problem was that the synchronization could be non-yet-finnshed when the test was running. So it was interfering with the result (the number of policy types for a ric was not the expected). 
Now the test waits until the RIC state is  "synchronized".

Patch-set: 2