Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I am not sure  that the URL is not correct
> 
> I Checked the logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/rtdv3-global-verify-master/2402/docs/offeredapis/redoc/api_policyManagementService.html
> 
> Something went wrong...
> Error downloading https://docs.onap.org/projects/onap-ccsdk-oran/en/latest/_downloads/1dbdc28e9b93a0ab56df1beca8f662be/swagger.json NetworkError when attempting to fetch resource.
> 
> Stack trace
> 
> s/<@https://cdn.jsdelivr.net/npm/redoc@2.0.0-alpha.17/bundles/redoc.standalone.js:40:15553
> e/</<@https://cdn.jsdelivr.net/npm/redoc@2.0.0-alpha.17/bundles/redoc.standalone.js:118:101808
> 
> ReDoc Version: 2.0.0-alpha.17
> Commit: de177a0

Thanks Eric for helping out with this!
I think the problem is due to the fact that the json file and the offeredapis.rst file doesn't exist in git yet, since they are committed for the first time in this commit. So my theory is that once this commit is in, the file will be generated at "https://docs.onap.org/projects/onap-ccsdk-oran/en/latest/_downloads" and the link will work. Do you think this might be correct?

Patch-set: 2