Update patch set 3
Patch Set 3:

A general question is why a lot of dependencies that are not used by the component as such should have to be defined here. Another general question is why we should have to define dependencies to internal springboot components. As I see it, ideally one dependency to the springboot framework (and the version of choice) should be enough.

Patch-set: 3
Reviewer: Gerrit User 5072 <5072@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>