Update path to SDNC rest interface in policy-management-service
The path to SDNC has been updated from  "/restconf/operations" to  "/rests/operations".
Hence, the update is needed in policy-management-service (PMS).
The path is made configurable so that the PMS can also work with older versions of SDNC.

The ODL password has also been updated in the latest version of SDNC.

Change-Id: If43c8632b4012437f9502fdac17ffa23f3a6719f
Issue-ID: CCSDK-3193
Signed-off-by: RehanRaza <muhammad.rehan.raza@est.tech>