Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

The concern I have here is that 'var' is new in Java 10.  This is fine for our ONAP compiles - we're already on Java 11.  However, if you know of any downstreams that are doing their own local compiles using earlier versions of Java (for example, because their IT department has not yet approved Java 10 or later), this change will break their compiles.

If the ORAN team is okay with that, that's fine - but I would suggest that we update our release notes to be explicit that downstreams doing local compiles should be aware that our code depends on Java 11 (technically Java 10 in this case, but better to just say Java 11 since that's what we're now testing against).

Patch-set: 2
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1